pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Jakub Mikłasz <jmiklasz21@gmail.com>"]
description = ""
name = "Manitobot"
version = "1.1.1"
[tool.poetry.dependencies]
"discord.py" = "^1.7.0"
flask = "^1.1"
python = "^3.8"
[tool.poetry.dev-dependencies]
